requirements.txt
requests==2.32.5
pandas==2.3.3
python-dotenv==1.2.1
duckdb==1.4.4
pytest==8.4.2
